requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-phonenumber-field[phonenumberslite]>=8.0
django-cors-headers>=4.3
gunicorn
psycopg2-binary
django-environ
django-storages[boto3]
boto3
celery[redis]
redis
drf-spectacular
pytest
pytest-django
factory-boy
drf_yasg